requirements.txt
# Core dependencies for Stock Market AI Agent
transformers>=4.30.0
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
requests>=2.31.0
click>=8.1.0
python-dotenv>=1.0.0
cachetools>=5.3.0
yfinance>=0.2.0
rich>=13.0.0

# Web UI dependencies
streamlit>=1.28.0
plotly>=5.17.0

# Testing dependencies
pytest>=7.4.0
hypothesis>=6.82.0
